requirements.txt
dnspython
bson
fastapi
uvicorn
gunicorn[standard]
python-dateutil
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-speech
google-cloud-texttospeech
googleapis-common-protos
websockets==15.0.1   #pip install --upgrade websockets
soundcard
numpy